feat: basic policy for post/comments/community

+++ backend/app/Http/Controllers/CommentController.php
@@ -31,6 +31,8 @@ class CommentController extends Controller
 
     public function destroy(Comment $comment)
     {
+        $this->authorize('destroy', $comment);
+
         $comment->delete();
         return response()->json(["message" => 'Comment deleted'], 200);
     }
